PMD: Add a new PMD called virtq PMD
The new PMD has features to communicate with virtio-net driver.
It means a guest kernel driver of virtio-net device and virtio-net PMD on a
guest OS can directly transfer data from/to the virtq PMD.
To use this PMD, a modified KVM is needed.

Both dpdk applications oh the host and guest can run as primary
process. So the applications can reboot independently.

While transferring data, if an one of appcalition is stopped, a virtual
connection between a virtq PMD and a viritio-net driver will be
disconnected automatically. And, the virtq PMD and the virtio-net driver
shows 'link down' status to the application.